// src/app/mejoramiento/aforos/models/AforoDato.ts
export interface AforoDato{
    idAforoDato: Number|null;
    idAforo:Number|null;
    fechaHora: String;
    automoviles: Number;
    busesVan: Number;
    busesBuseta: Number;
    busesSitpAlimentadores: Number;
    camionesC2p: Number;
    camionesC2g: Number;
    camionesC3C4: Number;
    camionesC5C5: Number;
    tpdParcial: Number;
   // auditoriausuario: String;
    //auditoriaFecha: Date|null;
  }